refactor: ArtworkRepository
